refactor: Data structure deck created, corrected collection
